+++ packages/terra-form-select/src/_SelectUtil.js
@@ -1,3 +1,4 @@
+import React from 'react';
 import Variants from './_constants';
 import MenuUtil from './_MenuUtil';
 
@@ -83,6 +84,31 @@ class SelectUtil {
   static valueDisplay(props, value) {
     const option = MenuUtil.findByValue(props.children, value);
     return option ? option.props.display : value;
+  }
+
+  /**
+   * Parses options in Select and return total number of options
+   * @param {React Children} children
+   * @return {number} - Number of total Options
+   */
+  static getTotalNumberOfOptions(children) {
+    let totalNumberOfOptions = 0;
+
+    React.Children.forEach(children, (child) => {
+      if (child.type.isOption) {
+        totalNumberOfOptions += 1;
+      }
+
+      if (child.type.isOptGroup) {
+        React.Children.forEach(child.props.children, (grandChild) => {
+          if (grandChild.type.isOption) {
+            totalNumberOfOptions += 1;
+          }
+        });
+      }
+    });
+
+    return totalNumberOfOptions;
   }
 }