+++ packages/terra-icon/src/icon/IconErrorLowLight.jsx
@@ -1,12 +1,20 @@
 /* eslint-disable */
 import React from 'react';
+import classNames from 'classnames/bind';
 import IconBase from '../IconBase';
+import styles from '../IconErrorLowLight.module.scss';
+
+const cx = classNames.bind(styles);
 
 const SvgIcon = (customProps) => {
   const attributes = Object.assign({}, customProps);
+  const iconClassNames = cx([
+    'IconErrorLowLight',
+    customProps.className,
+  ]);
 
   return (
-    <IconBase {...attributes}>
+    <IconBase className={iconClassNames} {...attributes}>
       <path fill="#F94E4E" d="M24 0c13.3 0 24 10.7 24 24S37.3 48 24 48 0 37.3 0 24C0 10.8 10.7 0 23.9 0h.1z" ></path><path fill="#B1B5B6" d="M7 21.5h34v5H7v-5z" ></path>
     </IconBase>
   );