+++ packages/terra-form-textarea/CHANGELOG.md
@@ -6,6 +6,11 @@ Unreleased
 ### Changed
 * Added 'disabled' prop to InputField component
 
+3.9.0 - (April 2, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 3.8.0 - (March 26, 2019)
 ------------------
 ### Changed